Replace atomic re-exports with standalone implementations
Atomic algorithm files now contain actual self-contained implementations
instead of re-exporting from progressive modules. Progressive files only
contain step-by-step visualization logic and import shared helpers from
their corresponding atomic modules.
